requirements.txt
pyaml
PyExifTool
